+++ src/main/java/org/khl/chat/service/MessageServiceImpl.java
@@ -13,7 +13,7 @@ import org.khl.chat.exception.NotFoundException;;
  @Service
 public class MessageServiceImpl implements MessageService{
 	
-	private static final Map<Integer, Message> MESSAGE_MAP_REPOSITORY = new HashMap<>();
+	private static final Map<Long, Message> MESSAGE_MAP_REPOSITORY = new HashMap<>();
 	private static final AtomicInteger MESSAGE_ID = new AtomicInteger();
 
 	
@@ -22,13 +22,13 @@ public class MessageServiceImpl implements MessageService{
 	
 	@Override
 	public void send(MessageDto msgDto) {
-		int msg_id = MESSAGE_ID.incrementAndGet();
+		Long msg_id = (long)MESSAGE_ID.incrementAndGet();
 		msgDto.setId(msg_id);
 		MESSAGE_MAP_REPOSITORY.put(msg_id, new Message(msgDto));
 	}
 
 	@Override
-	public void delete(int id) {
+	public void delete(Long id) {
 		if (MESSAGE_MAP_REPOSITORY.containsKey(id))
 			MESSAGE_MAP_REPOSITORY.remove(id);
 		else throw new NotFoundException("Сообщение не найдено");
@@ -36,7 +36,7 @@ public class MessageServiceImpl implements MessageService{
 	}
 
 	@Override
-	public MessageDto edit(int id, String text) {
+	public MessageDto edit(Long id, String text) {
 		if (MESSAGE_MAP_REPOSITORY.containsKey(id))	{	
 			Message msg = new Message();
 			msg.setValue(text);
